Renamed newly added GAN models
These names are more descriptive, and can also be parsed by the code to determine the latent vector size. However, the one with a latent vector of 32 does not work due to an error in Python, that seems to be caused by my PyTorch version. I've written code to work around this, but this seems to be another version of the same bug