Added computation budget to A*
A* was legitimately taking too long to fail on certain bad dungeons. The state space was large, because there were so many keys. For every possible number of keys that could be possessed, the agent was visiting every location in the map. However, the problem was that the triforce was in a separate disconnected component. A* was not frozen ... simply taking forever to search.

So, I added a budget after which point A* fails in a way that forces the CPPN to be more likely to place a room in a formerly unoccupied location ... in order to better connect rooms.

This allowed me to get through 4 generations of evolution before crashing with a new error.