Tweaking A* budget, but found other problem
I thought some levels seem to legitmately need a larger A* budget, but rather than try to pick the right busget a priori, I instead increase it by a tiny bit whenever the available amount proves insufficient.

But that didn't fix my problem either, so I commented this trick out. The problem here seems to be that the ladder item is needed to beat the level this fails on.