The GameGAN version of wox was old?
I replaced the wox code in GameGAN with the wox code in MMNEAT. Hopefully, this doesn't break anything, but it did fix something:

It is not possible to load/resume a terminated evolutionary run. More generally, it is possible to use Easy.load to load an ArrayList, among other things.