Changes to batch files and jar
@thehedgeify This change will probably require some updates to the Docker image.

As with the master branch, I've updated the Zelda code to use the new GAN model. I've tested it out, and I think we're still getting interesting results ... I'll save some in the preliminary results folder on Google Drive.

The other changes:
The Objective batch files now require a numeric command line parameter for determining the name of the save directory.

There is also a SeeResults.bat file that can be used to watch the results of any of the runs.