Refactored CPPN to GAN dungeon creation, but also found bug
Instead of having the CPPN make the inputs to the GAN directly, this is now handled by a general Builder interface. Currently, the only class that implements this interface uses a CPPN.

This code did not change, but the interactive evolution code is actually broken now ... however, I think this was caused by previous changes for the sake of objective evolution. Need to fix.