Attempting to generate a level-target fitness
The MarioLevelTask now has an objective that tries to load a Mario level in VGLC format and compare it to the evolved level.

However, because @thehedgeify enhanced the GAN encoding, the raw VGLC data that I added to the repo (VGLC\SuperMarioBros\) crashes the code when I attempt to load it. I could write some work-arounds for this in the code, but I suspect @thehedgeify has some altered data files that we could use instead.

Basically, there is a command in MarioLevelTask:
targetLevel = MarioLevelUtil.listLevelFromVGLCFile(levelFileName);
This command could be replaced with anything that somehow gets some data file into the List of Lists format.