CPPN outputs for down and right doors
Now the CPPN has outputs for whether there is a right door and whether there is a down door. Up and left are not needed, since the up/left doors in adjacent rooms are made when the right/down doors are made. This seems to work ok with the objective evolution, but A* is not able to successfully make these levels playable. Also, there is some kind of map display issue with the doors. Need to investigate more.