CPPN decides door type now
CPPN to GAN levels were inserting completely different door types each time a dungeon was created because of the use of random number generator. Now, the CPPN or the evolved GAN genotype provides numbers that are interpreted as different door types: regular, locked, soft locked, bombable

Can this fix the issue with A* making levels playable? There are probably still some levels that are impossible to beat because there is no door, but I should try